fix parameter ignore_target_role
Two things didn't worked the compare and the removal of the target-role
attribute.

The actual XML cib part was something like (example)

<clone id="clone_ping">
  <meta_attributes>
    <nvpair name="target-role" value="Started"/>
  </meta_attributes>
  <crmsh-ref id="res_ping"/>
</clone>

there old code was searching for attributes of the resource and not for
the meta_attributes child element.

I also change the is_same function to compare the actual cib xml object
and not the clean parser object. otherwise the ignore cleanup would be
useless